==> MACD Strategy US/3.19_MACD_trailingloss.Rmd <==

# Initialization
```{r}
rm(list=ls())
source("Utils.R")

# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
# MACD Strategy
# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
macd_strategy = function(dat_i) {

  # ------------------------------------------------------------------------------------------------
  # Parameters To Tune:
  # - trailing stop loss or fixed stop loss
  # - Use 200 EMA as filter                                                     (better without)
  # - % below pullback to set stop loss
  # - % to skip trade if stop loss X% is too large
  # - % to skip trade if price has already moved X% in the direction of trade
  # ------------------------------------------------------------------------------------------------
  
  # empty data.frame to return in case of errors
  empty_out = data.frame(entry.date=as.Date(NA),exit.date=as.Date(NA),outcome=NA,entry.price=NA,exit.price=NA)
  
  # return if data is less than 200 in length and cant compute 200ema
  if (nrow(dat_i) <= 200) return(empty_out)

  # techinicals
  ema12       = ema(dat_i$close,  12)
  ema26       = ema(dat_i$close,  26)
  ema200      = ema(dat_i$close,  200)
  macd12_26_9 = macd(dat_i$close, 12, 26, 9)

  # ema filter in ranging markets, returns when price is in a range
  #ema200_filter = ema_filter(ema200,dat_i$close,0,0.1)
  
  # entry
  macd_cross = crossover(macd12_26_9$signal,macd12_26_9$MACD)
  
  # entry signals, using ema filter improves nyse but not for nasdaq
  #macd_entry = which(macd_cross & macd12_26_9$MACD < 0)
  macd_entry = which(macd_cross & macd12_26_9$MACD < 0 & dat_i$close > ema200)
  #macd_entry = which(macd_cross & macd12_26_9$MACD < 0 & dat_i$close > ema200 & !ema200_filter)
  
  # loop through macd entries, how to find prior swing low, use that as stop loss, and take profit is 1.5x greater
  out_date_entry  = as.Date(c())
  out_date_exit   = as.Date(c())
  out_return      = as.numeric(c())
  out_price_entry = as.numeric(c())
  out_price_exit  = as.numeric(c())
  
  for (m in macd_entry) {
    
    # get local region of current price/date
    price     = dat_i$close[max(c((m - 100),1)):m]
    price_low = dat_i$low[max(c((m - 100),1)):m]
    date      = dat_i$date[max(c((m - 100),1)):m]
    
    # compute prior swing low
    price_smooth = smooth(c(price, price[length(price)]*seq(1.01,by=0.01,length=50)),0.1)[1:length(price)]
    minimas      = local_extrema(price_smooth,max=FALSE)
    min_ind      = minimas[length(minimas)]
    min_date     = date[min_ind]
    
    # find minimum in local region of min_date
    min_date = date[which.min(price_low[max(c(1,(min_ind - 5))):min(c(length(price_low),(min_ind + 5)))]) + min_ind - 6]
    
    # derive the exact price of crossover, and thus, entry price
    price_entry = 351/28*(28/351* dat_i$close[m-1] + 11/13*(ema12[m-2] - ema12[m-1]) + 25/27*(ema26[m-1] - ema26[m-2]))
    
    # set stop price
    price_stop   = dat_i$low[dat_i$date == min_date]*0.99 # put stop loss 1.0% below prior swing low
    
    # skip if stop loss is too large (5.0%)
    crit1 = abs(price_stop - price_entry)/price_entry*100 > 5
    if (ifelse(is.na(crit1),FALSE,crit1)) next
    
    # skip if price has already moved by over 2.5% in favor of the trade before the entry ()
    crit2 = (dat_i$close[m] - dat_i$close[m-3])/dat_i$close[m]*100 > 2.5
    if (ifelse(is.na(crit2),FALSE,crit2)) next

    # trailing stop loss
    stop_per   = (price_stop - price_entry)/price_entry*100
    max_vals   = cummax(dat_i$high[m:(nrow(dat_i)-1)])
    draw_downs = c(0,(dat_i$low[(m+1):nrow(dat_i)] - max_vals)/max_vals*100)
    ind_matchs = which(draw_downs <= stop_per)
    exit_date  = min(ind_matchs + m - 1)

    # store the trade if there is an exit, and if the exit date is not in the future
    if (price_entry >= dat_i$low[m] & price_entry <= dat_i$high[m])  {
      if (is.finite(exit_date) & !is.na(dat_i$date[exit_date])) {
        
        # compute exit price
        price_exit = max(c(dat_i$high[m:(exit_date-1)],dat_i$open[exit_date]))*(100+stop_per)/100
        if (price_exit > dat_i$open[exit_date]) price_exit = dat_i$open[exit_date]
        
        # compute return
        return     = (price_exit - price_entry)/price_entry*100
  
        # set dates
        date_entry = dat_i$date[m]
        date_exit  = dat_i$date[exit_date]
        
        # save data
        out_date_entry  = c(out_date_entry,date_entry)
        out_date_exit   = c(out_date_exit,date_exit)
        out_return      = c(out_return, return)
        out_price_entry = c(out_price_entry,price_entry)
        out_price_exit  = c(out_price_exit,price_exit)
        
      }      
    }
    
  }
  
  # output data frame with all trades
  dat_out = data.frame(entry.date  = as.Date(out_date_entry),
                       exit.date   = as.Date(out_date_exit),
                       outcome     = as.numeric(out_return),
                       entry.price = as.numeric(out_price_entry),
                       exit.price  = as.numeric(out_price_exit))
  
  # if no trades return empty matrix
  if (nrow(dat_out)==0) {
    return(empty_out)
  } else {
    return(dat_out)
  }
  
}
```

# Load and Run MACD strategy
```{r}
exchange = "nasdaq" # nasdaq nyse nysearca

# set up stocks
stocks = read.csv(paste0(toupper(exchange),".csv"))
stocks$macd_return = NA
stocks$macd_n      = NA
stocks$gradient    = NA
stocks$residual    = NA

# dataframe containing history of all trades
trade_hist = data.frame(ticker=character(),
                        entry.date=as.Date(character()),exit.date=as.Date(character()),
                        outcome=as.numeric(),
                        entry.price=as.numeric(),exit.price=as.numeric())

# read stock data
for (i in 1:nrow(stocks)) {
  
    print(stocks$Symbol[i])
  
    # load stock
    path = file.path(Sys.getenv("USERPROFILE"),"Desktop",paste0(exchange," stocks"))
    df   = read.csv(paste0(path,"/",stocks$Symbol[i],".csv"),sep=",",header=TRUE,stringsAsFactors=FALSE)
    
    # run strategies
    macd_out = macd_strategy(df)
    
    # appending trade history
    trade_hist = rbind(trade_hist, cbind(ticker=stocks$Symbol[i],macd_out))
    
    # save winning percentages and number of trades
    macd_out = macd_out[!is.na(macd_out$entry.date),] # remove NA rows
    
    stocks$macd_return[i] = sum(macd_out$outcome>0)/nrow(macd_out)
    stocks$macd_n[i]      = nrow(macd_out)
    
    # compute linear regression
    LR = lm(1:nrow(df) ~ df$close)
    
    stocks$gradient[i] = LR$coefficients[2]
    stocks$residual[i] = mean(LR$residuals^2)
    
}

# compute win rate and average position length
trade_hist = trade_hist[!is.na(trade_hist$outcome),]

sprintf("Win Rate: %.1f%%", sum(trade_hist$outcome>0)/nrow(trade_hist)*100)
sprintf("Risk Reward Ratio: %.1f",abs(mean(trade_hist$outcome[trade_hist$outcome > 0])/mean(trade_hist$outcome[trade_hist$outcome < 0])))
sprintf("Average Hold Length: %.0f days", mean(trade_hist$exit.date-trade_hist$entry.date))
```

# Compute Return with Non-Overlapping Trades (Realistic Metrics)
```{r}
# sort data
trade_hist = trade_hist[order(trade_hist$entry.date),]

# filter data
trade_hist1 = trade_hist
trade_hist1 = trade_hist1[trade_hist1$ticker %in% stocks$Symbol[!is.na(stocks$Cap) & stocks$macd_n != 0],]
trade_hist1 = trade_hist1[trade_hist1$ticker %in% stocks$Symbol[stocks$Cap>1000],]
#trade_hist1 = trade_hist1[trade_hist1$ticker %in% stocks$Symbol[stocks$gradient>0],]
#trade_hist1 = trade_hist1[trade_hist1$ticker %in% stocks$residual<750000,]

# initialize counters
rm_rows = c()
last_exit = trade_hist1$exit.date[1]

# loop through all data
for (i in 1:nrow(trade_hist1)) {
  if (trade_hist1$entry.date[i] <= last_exit) {
    rm_rows = c(rm_rows,i)
  } else {
    last_exit = trade_hist1$exit.date[i]
  }
}

# filter out the removed rows
trade_hist1 = trade_hist1[-rm_rows,]
row.names(trade_hist1) = NULL

# calculate length of trade
trade_hist1$trade_len  = trade_hist1$exit.date - trade_hist1$entry.date

# generate summarize statistics
rle          = rle(trade_hist1$outcome>0)
trade_period = as.numeric(trade_hist1$exit.date[nrow(trade_hist1)] - trade_hist1$entry.date[1])
cumreturn    = tail(cumprod(c(1,1+trade_hist1$outcome/100))-1,1)

p1 = list()
p1[[1]] = sprintf("----------------------- %s",exchange)
p1[[2]] = sprintf("Win Rate: %.1f%%", sum(trade_hist1$outcome>0)/nrow(trade_hist1)*100)
p1[[3]] = sprintf("Average Win:   %0.1f%% (%.0f Days)",
                  mean(trade_hist1$outcome[trade_hist1$outcome>0]),mean(trade_hist1$trade_len[trade_hist1$outcome>0]))
p1[[4]] = sprintf("Average Loss: %0.1f%% (%.0f Days)",
                  mean(trade_hist1$outcome[trade_hist1$outcome<=0]),mean(trade_hist1$trade_len[trade_hist1$outcome<=0]))
p1[[5]] = sprintf("Win vs Lose Streak: %i vs %i", 
                  max(rle$lengths[rle$values],na.rm=TRUE),max(rle$lengths[!rle$values],na.rm=TRUE))
p1[[6]] = sprintf("Cumulative Returns: %0.1f%% (%.1f Yrs, %.1f Trades/Yr)",
                  cumreturn*100,trade_period/365.25,nrow(trade_hist1)/(trade_period/365.25))
p1[[7]] = sprintf("Annualized Return:  %0.1f%%", (cumreturn^(365.25/trade_period)-1)*100)

# generate portfolio statistics
macd_trade = c(1,cumprod(c(1,1+trade_hist1$outcome/100))[-1])
drawdowns  = sapply(2:length(macd_trade), function(x) (macd_trade[x] - max(macd_trade[1:(x-1)]))/max(macd_trade[1:(x-1)])*100)
rle        = rle(drawdowns<0)

max_rle = which(rle$length == max(rle$lengths[rle$values==TRUE]))
start   = sum(rle$lengths[1:(max_rle-1)])
end     = start+rle$lengths[tail(max_rle,1)]

p2 = list()
p2[[1]] = sprintf("----------------------- Portfolio")
p2[[2]] = sprintf("Maximum Drawdown: %.1f%%", min(drawdowns))
p2[[3]] = sprintf("Longest Down Time: %0.1f Yrs (%s to %s)", 
                  as.numeric((trade_hist1$exit.date[end] - trade_hist1$entry.date[start])/365.25),
                  as.character(trade_hist1$entry.date[start]),as.character(trade_hist1$exit.date[end]))

# projected return over entire history
plot(as.Date(trade_hist1$entry.date),macd_trade[-1],type="l",xlab="Date",ylab="Portfolio",log="y")
abline(h=c(1:10,2:10*10,2:10*100,2:10*1000,2:10*10000,2:10*100000),col=alpha("grey",0.5))
abline(v=trade_hist1$entry.date[start],col="red")
abline(v=trade_hist1$exit.date[end],col="red")
points(trade_hist1$entry.date[which.min(drawdowns)],macd_trade[which.min(drawdowns)+1],col="deeppink")

# display report
cat(paste0(unlist(p1),collase="\n"))
cat(paste0(unlist(p2),collase="\n"))
```

# Save Trades
```{r}
# plot and save all trades
for (i in 1:nrow(trade_hist1)) {

  # load stock
  path    = file.path(Sys.getenv("USERPROFILE"),"Desktop",paste0(exchange," stocks"))
  df      = read.csv(paste0(path,"/",trade_hist1$ticker[i],".csv"),sep=",",header=TRUE,stringsAsFactors=FALSE)
  df$date = as.Date(df$date)
  
  # crop df
  ema200 = ema(df$close,200)
  ema200 = ema200[df$date > (trade_hist1$entry.date[i] - 25) & df$date < (trade_hist1$exit.date[i] + 10)]
  df     = df[df$date > (trade_hist1$entry.date[i] - 25) & df$date < (trade_hist1$exit.date[i] + 10),]
  
  # plot
  savepath = file.path(Sys.getenv("USERPROFILE"),"Desktop",paste0(exchange," trades"))
  plot_stock(df,emas=list(ema200),
             entry=trade_hist1$entry.date[i],exit=trade_hist1$exit.date[i],
             entry.price=trade_hist1$entry.price[i],exit.price=trade_hist1$exit.price[i],
             save=TRUE,savename=paste0(savepath,"/",i,"_",trade_hist1$ticker[i]," (",
                                       as.character(round(trade_hist1$outcome[i],1)),"%%)"))
  
}
```

# Find Next Trade
```{r}
exchange = "nasdaq" # nasdaq nyse nysearca

# load stocks database
stocks = read.csv(paste0(toupper(exchange),".csv"))
stocks$diffMACD    = NA
stocks$price.entry = NA
stocks$per.stop    = NA

# get list of tickers
for (i in 1:length(stocks$Symbol)) {
  
    print(stocks$Symbol[i])

    # load stock
    path = file.path(Sys.getenv("USERPROFILE"),"Desktop",paste0(exchange," stocks"))
    df   = read.csv(paste0(path,"/",stocks$Symbol[i],".csv"),sep=",",header=TRUE,stringsAsFactors=FALSE)
    if (nrow(df) < 200) next
    
    # macd technicals & compute how close macd is to crossover
    macd12_26_9 = macd(df$close, 12, 26, 9)
  
    macd_n = tail(macd12_26_9$MACD,1)
    sig_n  = tail(macd12_26_9$signal,1)
    if (macd_n > sig_n | sig_n  >= 0 | macd_n >= 0) next
    diffnorm = abs((macd_n - sig_n)/macd_n)
    
    # emas techinicals
    ema12       = ema(df$close, 12)
    ema26       = ema(df$close, 26)
    ema200      = ema(df$close, 200)
    
    #if (ema200[nrow(df)] > df$close[nrow(df)]) next
    
    # analyze price action at latest price
    m = nrow(df)

    # derive the exact price of crossover, and thus, entry price
    price_entry = 351/28*(28/351* df$close[m-1] + 11/13*(ema12[m-2] - ema12[m-1]) + 25/27*(ema26[m-1] - ema26[m-2]))
    
    # get local region of current price/date
    price     = df$close[max(c((m - 100),1)):m]
    price_low = df$low[max(c((m - 100),1)):m]
    date      = df$date[max(c((m - 100),1)):m]
    
    # compute prior swing low
    price_smooth = smooth(c(price, price[length(price)]*seq(1.01,by=0.01,length=50)),0.1)[1:length(price)]
    minimas      = local_extrema(price_smooth,max=FALSE)
    min_ind      = minimas[length(minimas)]
    min_date     = date[min_ind]
    
    # find minimum in local region of min_date
    min_date = date[which.min(price_low[max(c(1,(min_ind - 5))):min(c(length(price_low),(min_ind + 5)))]) + min_ind - 6]

    # set stop price
    price_stop = df$low[df$date == min_date]*0.99 # put stop loss 1.0% below prior swing low
    stop_per   = (price_stop - price_entry)/price_entry*100
    
    # skip if stop loss is too large (5.0%)
    crit1 = abs(price_stop - price_entry)/price_entry*100 > 5
    if (ifelse(is.na(crit1),FALSE,crit1)) next
    
    # skip if price has already moved by over 2.5% in favor of the trade before the entry ()
    crit2 = (df$close[m] - df$close[m-3])/df$close[m]*100 > 2.5
    if (ifelse(is.na(crit2),FALSE,crit2)) next
    
    # save values
    stocks$diffMACD[i]    = diffnorm
    stocks$price.entry[i] = price_entry
    stocks$per.stop[i]    = stop_per
    
}

# Filter Stocks
stocks_enter = stocks
stocks_enter = stocks_enter[stocks_enter$Cap > 1000,]
stocks_enter = stocks_enter[!is.na(stocks_enter$diffMACD),]
stocks_enter = stocks_enter[order(stocks_enter$diffMACD,decreasing=FALSE),]
rownames(stocks_enter) = NULL

# Plot Stocks
for (i in 1:nrow(stocks_enter)) {
  
  # load stock
  path    = file.path(Sys.getenv("USERPROFILE"),"Desktop",paste0(exchange," stocks"))
  df      = read.csv(paste0(path,"/",stocks_enter$Symbol[i],".csv"),sep=",",header=TRUE,stringsAsFactors=FALSE)
  df$date = as.Date(df$date)
    
  # compute macd
  ema200 = ema(df$close,200)
  macd12_26_9 = macd(df$close, 12, 26, 9)
  
  # crop df
  ema200             = ema200[(nrow(df) - 66):nrow(df)]
  macd12_26_9$MACD   = macd12_26_9$MACD[(nrow(df) - 66):nrow(df)]
  macd12_26_9$signal = macd12_26_9$signal[(nrow(df) - 66):nrow(df)]
  df                 = df[(nrow(df) - 66):nrow(df),]
  
  # save
  savepath = file.path(Sys.getenv("USERPROFILE"),"Desktop","Next Trades")
  plot_stock(df,sub="macd",
             emas=list(ema200),macd=macd12_26_9,
             save=TRUE,savename=paste0(savepath,"/",i,"_",stocks_enter$Symbol[i]))
  
}
```

> Strategy 1: MACD
  - display all current stocks with MACD strategy close to giving entry (with price action 66 days + MACD), in order of diffMACD
  
  - validate MACD back-compute by testing samples
  - go through 100 trades and validate with tradingview to see that they are actually correct
  
  - algorithm to determine which trades to take when different tickers show signal on the same day



> Strategy 2: Trade after large moves
   - generate plots to show what happens after a large move in a ticker (seperated by market caps)
   - enter after a very large move in the market, and enter at the pull back of the large move,
     when candle is bullish (first green to close above red)
