==> MACD Strategy US/3.33_MACD_trailingloss_all_US_predict_MACD_X.Rmd <==

# Notes:
# - Using top 20 most likely MACD_diff produces the best annual return
# - Winning % does not differ amongst the 1st to 20th most likely values
# - ~40% of trades actually have MACD cross, the crossed trades have ~80% win, the non-crossed have ~40%

# Pre-Compute MACD Stats
```{r}
rm(list=ls())
source("Utils.R")

# set up stocks
nyse_stocks   = read.csv("NYSE.csv")
nasdaq_stocks = read.csv("NASDAQ.csv")
exchange      = c(rep("NYSE",nrow(nyse_stocks)),rep("NASDAQ",nrow(nasdaq_stocks)))

stocks = rbind(nyse_stocks[c("Symbol","Cap","Float")],nasdaq_stocks[c("Symbol","Cap","Float")])
stocks$exchange = exchange

# store values
macd_mat = array(rep(NA,nrow(stocks)*8000*3),dim=c(nrow(stocks),8000,3))

# read stock data
for (i in 1:nrow(stocks)) {
  
    print(stocks$Symbol[i])
  
    # check if stock data exists
    if (sum(paste0(stocks$Symbol[i],".csv") == 
        list.files(file.path(Sys.getenv("USERPROFILE"),"Desktop",paste0(stocks$exchange[i]," stocks")))) == 0) next
  
    # load stock
    path = file.path(Sys.getenv("USERPROFILE"),"Desktop",paste0(stocks$exchange[i]," stocks"))
    df   = read.csv(paste0(path,"/",stocks$Symbol[i],".csv"),sep=",",header=TRUE,stringsAsFactors=FALSE)
    
    # compute ema and macd entries
    macd12_26_9 = macd(df$Close, 12, 26, 9)
    macd_cross  = crossover(macd12_26_9$signal,macd12_26_9$MACD) & macd12_26_9$MACD < 0
    
    # store values
    macd_mat[i, (8000 - nrow(df) + 1):8000,1] = macd_cross
    macd_mat[i, (8000 - nrow(df) + 1):8000,2] = macd12_26_9$signal
    macd_mat[i, (8000 - nrow(df) + 1):8000,3] = macd12_26_9$MACD

}

# compute difference in MACD line and signal line
MACD_diff = abs((macd_mat[,,2] - macd_mat[,,3])/macd_mat[,,2])
MACD_diff[macd_mat[,,2] >= 0] = NA
MACD_diff[macd_mat[,,3] >= 0] = NA
MACD_diff[macd_mat[,,2] <= macd_mat[,,3]] = NA
MACD_diff[!is.finite(MACD_diff)] = NA
MACD_diff[stocks$Cap < 1000,] = NA

# change MACD diff values into rankings based on smallest number
for (i in 1:dim(MACD_diff)[2]) {
  if (sum(!is.na(MACD_diff[,i])) > 20) {
    MACD_diff[,i] = rank(MACD_diff[,i],ties.method="first")
  } else {
    MACD_diff[,i] = NA
  }
}

# only use top N
MACD_diff[MACD_diff>20] = NA
```

# Initialization
```{r}
# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
# MACD Strategy
# # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
macd_strategy = function(dat_i,ind_i) {

  # ------------------------------------------------------------------------------------------------
  # Parameters To Tune:
  # - special back-computed MACD entry                   (big deal, but need to think more about execution aespect)
  # - % to skip trade if stop loss X% is too large       (2% is best)
  # - Use prior Close/Low to identify min of pullback    (Close vs Low, Low is better)
  # - Use prior Close/Low of pullback for stop loss      (Close vs Low, Low is better)
  # - Use High/Close for retracement from top to exit    (High vs Close, High is better)
  # ------------------------------------------------------------------------------------------------
  
  # empty data.frame to return in case of errors
  empty_out = data.frame(entry.date=as.Date(NA),exit.date=as.Date(NA),outcome=NA,entry.price=NA,exit.price=NA,stop.percent=NA,
                         macd.order=NA,macd.cross=NA)
  
  # return if data is less than 35 in length and cant compute MACD
  if (nrow(dat_i) <= 35) return(empty_out)

  # techinicals
  ema12       = ema(dat_i$Close,  12)
  ema26       = ema(dat_i$Close,  26)
  macd12_26_9 = macd(dat_i$Close, 12, 26, 9)
 
  # entry
  macd_entry = unname(which(!is.na(MACD_diff[ind_i,(8000 - nrow(dat_i) + 1):8000]))) + 1
  macd_entry = macd_entry[macd_entry > 35]
  macd_entry = macd_entry[macd_entry < nrow(dat_i)]

  # initialize vectors for output
  out_date_entry   = as.Date(c())
  out_date_exit    = as.Date(c())
  out_return       = as.numeric(c())
  out_price_entry  = as.numeric(c())
  out_price_exit   = as.numeric(c())
  out_stop_percent = as.numeric(c())
  out_diff_order   = as.numeric(c())
  out_macd_cross   = as.numeric(c())
  
  for (m in macd_entry) {
    
    # get local region of current price/date
    price     = dat_i$Close[max(c((m - 100),1)):m]
    price_low = dat_i$Low[max(c((m - 100),1)):m]
    date      = dat_i$Date[max(c((m - 100),1)):m]
    
    # compute prior swing low
    price_smooth = smooth(c(price, price[length(price)]*seq(1.01,by=0.01,length=50)),0.1)[1:length(price)]
    minimas      = local_extrema(price_smooth,max=FALSE)
    min_ind      = minimas[length(minimas)]
    min_date     = date[min_ind]
    
    # find minimum in local region of min_date
    min_date = date[which.min(price_low[max(c(1,(min_ind - 5))):min(c(length(price_low),(min_ind + 5)))]) + min_ind - 6]
    
    # derive the exact price of crossover, and thus, entry price
    price_entry = 351/28*(28/351*dat_i$Close[m-1] + 11/13*(ema12[m-2] - ema12[m-1]) + 25/27*(ema26[m-1] - ema26[m-2]))
    
    # if price entry is not hit in the next day intraday, then skip trade
    # if using limit order, and price_entry is above opening price, then enter at open
    if (!(price_entry >= dat_i$Low[m] & price_entry <= dat_i$High[m])) next
    if (price_entry > dat_i$Open[m]) price_entry = dat_i$Open[m]
    
    # set stop price and stop loss percentage
    price_stop = dat_i$Low[dat_i$Date == min_date][1] # Use Low or Close
    stop_per   = (price_stop - price_entry)/price_entry*100

    # skip if stop loss is too large, or if its positive which means the code fked up
    if (stop_per < -2 | stop_per > 0) next
  
    # trailing stop loss
    max_vals   = cummax(dat_i$High[m:(nrow(dat_i)-1)])
    draw_downs = c(9999999,(dat_i$Low[(m+1):nrow(dat_i)] - max_vals)/max_vals*100)
    ind_matchs = which(draw_downs <= stop_per)
    exit_date  = min(ind_matchs + m - 1)
  
    # if there is no valid exit date, then skip trade
    if (!(is.finite(exit_date) & !is.na(dat_i$Date[exit_date]))) next

    # compute exit price
    price_exit = max(dat_i$High[m:(exit_date-1)])*(100+stop_per)/100
    if (price_exit > dat_i$Open[exit_date]) price_exit = dat_i$Open[exit_date]
    
    # compute return
    return = (price_exit - price_entry)/price_entry*100

    # set dates
    date_entry = dat_i$Date[m]
    date_exit  = dat_i$Date[exit_date]
    
    # save data
    out_date_entry   = c(out_date_entry,date_entry)
    out_date_exit    = c(out_date_exit,date_exit)
    out_return       = c(out_return,return)
    out_price_entry  = c(out_price_entry,price_entry)
    out_price_exit   = c(out_price_exit,price_exit)
    out_stop_percent = c(out_stop_percent,stop_per)
    
    out_diff_order = c(out_diff_order,MACD_diff[ind_i,][m + (8000 - nrow(dat_i)) - 1])
    out_macd_cross = c(out_macd_cross,macd12_26_9$signal[m-1] > macd12_26_9$MACD[m-1] & macd12_26_9$signal[m] < macd12_26_9$MACD[m])
    
  }

  # output data frame with all trades
  dat_out = data.frame(entry.date   = as.Date(out_date_entry),
                       exit.date    = as.Date(out_date_exit),
                       outcome      = as.numeric(out_return),
                       entry.price  = as.numeric(out_price_entry),
                       exit.price   = as.numeric(out_price_exit),
                       stop.percent = as.numeric(out_stop_percent),
                       macd.order   = as.numeric(out_diff_order),
                       macd.cross   = as.numeric(out_macd_cross))
  
  # if no trades return empty matrix
  if (nrow(dat_out)==0) {
    return(empty_out)
  } else {
    return(dat_out)
  }
  
}
```

# Load and Run MACD strategy on all US stocks
```{r}
# set up stocks
nyse_stocks   = read.csv("NYSE.csv")
nasdaq_stocks = read.csv("NASDAQ.csv")
exchange      = c(rep("NYSE",nrow(nyse_stocks)),rep("NASDAQ",nrow(nasdaq_stocks)))

stocks = rbind(nyse_stocks[c("Symbol","Cap","Float")],nasdaq_stocks[c("Symbol","Cap","Float")])
stocks$exchange = exchange

# dataframe containing history of all trades
trade_hist = data.frame(ticker=character(),exchange=character(),
                        entry.date=as.Date(character()),exit.date=as.Date(character()),
                        outcome=as.numeric(),
                        entry.price=as.numeric(),exit.price=as.numeric(),stop.percent=as.numeric(),
                        
                        macd.order=as.numeric(),macd.cross=as.numeric())

# read stock data
for (i in 1:nrow(stocks)) {
  
    print(stocks$Symbol[i])
  
    # check if stock data exists
    if (sum(paste0(stocks$Symbol[i],".csv") == 
          list.files(file.path(Sys.getenv("USERPROFILE"),"Desktop",paste0(stocks$exchange[i]," stocks")))) == 0) next
  
    # load stock
    path = file.path(Sys.getenv("USERPROFILE"),"Desktop",paste0(exchange[i]," stocks"))
    df   = read.csv(paste0(path,"/",stocks$Symbol[i],".csv"),sep=",",header=TRUE,stringsAsFactors=FALSE)

    # run strategies
    macd_out = macd_strategy(df,i)
    
    # appending trade history
    trade_hist = rbind(trade_hist, cbind(ticker=stocks$Symbol[i],exchange=exchange[i],macd_out))
    
}

trade_hist = trade_hist[!is.na(trade_hist$outcome),]

# compute win rate and average position length
sprintf("Win Rate: %.1f%%",sum(trade_hist$outcome>0)/nrow(trade_hist)*100)
sprintf("Risk Reward Ratio: %.1f",abs(mean(trade_hist$outcome[trade_hist$outcome > 0])/mean(trade_hist$outcome[trade_hist$outcome < 0])))
sprintf("Average Hold Length: %.0f days", mean(trade_hist$exit.date-trade_hist$entry.date))

# save trade history
#write.csv(trade_hist,file="trade_hist.csv",row.names=FALSE)

# produce more statistics for the MACD_diff
sprintf("Proportion of Actual MACD Cross: %.1f%%",sum(trade_hist$macd.cross)/nrow(trade_hist)*100)
sprintf("Winning %% for MACD Crosses: %.1f%%",sum(trade_hist$outcome[trade_hist$macd.cross==1]>0)/sum(trade_hist$macd.cross==1)*100)
sprintf("Winning %% for non MACD Crosses: %.1f%%",sum(trade_hist$outcome[trade_hist$macd.cross==0]>0)/sum(trade_hist$macd.cross==0)*100)

#for (i in 1:20) {
#  print(sprintf("Winning %% for MACD rank %.0f: %.1f%%",
#                i,sum(trade_hist$outcome[trade_hist$macd.order==i]>0)/sum(trade_hist$macd.order==i)*100))
#}
```

# (Realistic Metrics) Compute Return with Non-Overlapping Trades 
```{r}
# sort data
trade_hist = trade_hist[order(trade_hist$entry.date),]

# filter data
trade_hist1 = trade_hist
trade_hist1 = trade_hist1[trade_hist1$ticker %in% stocks$Symbol[!is.na(stocks$Cap)],]
trade_hist1 = trade_hist1[trade_hist1$ticker %in% stocks$Symbol[stocks$Cap>1000],]

# initialize counters
rm_rows = c()
last_exit = trade_hist1$exit.date[1]

# loop through all data
for (i in 1:nrow(trade_hist1)) {
  if (trade_hist1$entry.date[i] <= last_exit) {
    rm_rows = c(rm_rows,i)
  } else {
    last_exit = trade_hist1$exit.date[i]
  }
}

# filter out the removed rows
trade_hist1 = trade_hist1[-rm_rows,]
row.names(trade_hist1) = NULL

# calculate length of trade
trade_hist1$trade_len  = trade_hist1$exit.date - trade_hist1$entry.date

# generate summarize statistics
rle          = rle(trade_hist1$outcome>0)
trade_period = as.numeric(trade_hist1$exit.date[nrow(trade_hist1)] - trade_hist1$entry.date[1])
cumreturn    = tail(cumprod(c(1,1+trade_hist1$outcome/100))-1,1)
n_trades_yr  = nrow(trade_hist1)/(trade_period/365.25)

p1 = list()
p1[[1]] = sprintf("----------------------- nyse + nasdaq")
p1[[2]] = sprintf("Win Rate: %.1f%%", sum(trade_hist1$outcome>0)/nrow(trade_hist1)*100)
p1[[3]] = sprintf("Average Win:   %0.1f%% (%.0f Days)",
                  mean(trade_hist1$outcome[trade_hist1$outcome>0]),mean(trade_hist1$trade_len[trade_hist1$outcome>0]))
p1[[4]] = sprintf("Average Loss: %0.1f%% (%.0f Days)",
                  mean(trade_hist1$outcome[trade_hist1$outcome<=0]),mean(trade_hist1$trade_len[trade_hist1$outcome<=0]))
p1[[5]] = sprintf("Win vs Lose Streak: %i vs %i", 
                  max(rle$lengths[rle$values],na.rm=TRUE),max(rle$lengths[!rle$values],na.rm=TRUE))
p1[[6]] = sprintf("Cumulative Returns: %0.1f%% (%.1f Yrs, %.1f Trades/Yr)",
                  cumreturn*100,trade_period/365.25,n_trades_yr)
p1[[7]] = sprintf("Annualized Return:  %0.1f%%", (cumreturn^(365.25/trade_period)-1)*100)

# generate portfolio statistics
macd_trade = cumprod(c(1,1+trade_hist1$outcome/100))[-1]
drawdowns  = sapply(1:length(macd_trade), function(x) (macd_trade[x] - max(macd_trade[1:(x-1)]))/max(macd_trade[1:(x-1)])*100)
rle        = rle(drawdowns<0)

max_rle = which(rle$length == max(rle$lengths[rle$values==TRUE]))
start   = sum(rle$lengths[1:(max_rle-1)])
end     = start+rle$lengths[tail(max_rle,1)]

p2 = list()
p2[[1]] = sprintf("----------------------- Portfolio")
p2[[2]] = sprintf("Maximum Drawdown: %.1f%%", min(drawdowns))
p2[[3]] = sprintf("Longest Down Time: %0.1f Yrs (%s to %s)", 
                  as.numeric((trade_hist1$exit.date[end] - trade_hist1$entry.date[start])/365.25),
                  as.character(trade_hist1$entry.date[start]),as.character(trade_hist1$exit.date[end]))
p2[[4]] = sprintf("Proportion of Days in Position: %0.1f%%",
                  as.numeric(sum(trade_hist1$exit.date - trade_hist1$entry.date)/
                    as.numeric((tail(trade_hist1$exit.date,1) - trade_hist1$entry.date[1])))*100)

# projected return over entire history
par(mfrow=c(2,1))
plot(as.Date(trade_hist1$entry.date),macd_trade,type="l",xlab="",ylab="Portfolio",log="y")
abline(h=10^(0:10),col=alpha("grey",0.33))
abline(v=as.Date(paste0(1990:2030,"-01-01")),col=alpha("grey",0.5))
points(trade_hist1$entry.date[which.min(drawdowns)],macd_trade[which.min(drawdowns)+1],col="deeppink")

osc = sapply(1:length(macd_trade),function(x) ifelse(x <= round(n_trades_yr,0),
                                                    NA,
                                                    (macd_trade[x] - macd_trade[x-round(n_trades_yr,0)])/macd_trade[x]))*100
plot(as.Date(trade_hist1$entry.date),type="l",osc,col="deeppink",xlab="Date",ylab="Annual Return")
abline(h=seq(0,100,by=10),col=alpha("grey",0.5))
abline(v=as.Date(paste0(1990:2030,"-01-01")),col=alpha("grey",0.5))

# display report
cat(paste0(unlist(p1),collapse="\n"))
cat(paste0(unlist(p2),collapse="\n"))

# save trade history
write.csv(trade_hist1,"trade_hist_real_return.csv",row.names=FALSE)
```

# Plot Annual Returns
```{r}
comission  = 0.1 # more like 0.025
macd_trade = cumprod(c(1,1+(trade_hist1$outcome - comission)/100))[-1]

# load sp500 data
sp500       = read.csv("old code/SPY.csv")
sp500$Date  = as.Date(sp500$Date)
sp500$Close = as.numeric(sp500$Close)

# load nasdaq composite index data
nasdaq       = read.csv("old code/IXIC.csv")
nasdaq$Date  = as.Date(nasdaq$Date)
nasdaq$Close = as.numeric(nasdaq$Close)

# initialize vectors
return_sp500  = c()
return_nasdaq = c()
return_macd   = c()

# date range
date_range = 1993:2021

# loop through each year
for (d in date_range) {
  
  # set date range (inclusive)
  start_date = as.Date(paste0(d,"-01-01"))
  end_date   = as.Date(paste0(d,"-12-31"))
  
  # filter
  sp500_1yr  = sp500$Close[sp500$Date >= start_date & sp500$Date <= end_date]
  nasdaq_1yr = nasdaq$Close[nasdaq$Date >= start_date & nasdaq$Date <= end_date]
  macd_1yr   = macd_trade[trade_hist1$entry.date >= start_date & trade_hist1$exit.date <= end_date]
  
  # 1 year return
  return_sp500_1yr  = (tail(sp500_1yr,1) - sp500_1yr[1])/sp500_1yr[1]*100
  return_nasdaq_1yr = (tail(nasdaq_1yr,1) - nasdaq_1yr[1])/nasdaq_1yr[1]*100
  return_macd_1yr   = (tail(macd_1yr,1) - macd_1yr[1])/macd_1yr[1]*100
  
  # append data
  return_sp500  = c(return_sp500,return_sp500_1yr)
  return_nasdaq = c(return_nasdaq,return_nasdaq_1yr)
  return_macd   = c(return_macd, ifelse(length(return_macd_1yr) == 0,0,return_macd_1yr))
  
}

# plot
df_plot  = t(data.frame(SP500 = return_sp500,NASDAQ = return_nasdaq, MACD = return_macd))
colnames(df_plot) = date_range

barplot(df_plot,col = c("deepskyblue","tomato","darkolivegreen2"), beside = TRUE,ylab="Annual Return (%)",xlab="Year")
abline(h=seq(-50,10000,by=50),col=alpha("grey",0.5))
legend("topright",c("SP500","NASDAQ","STRATEGY"),col=c("deepskyblue","tomato","darkolivegreen2"),lty=1,lwd=10)

# print report
p = list()
p[[1]] = sprintf("Annualized Returns")
p[[2]] = sprintf("SP500 Return    %0.1f%%",mean(return_sp500))
p[[3]] = sprintf("NASDAQ Return   %0.1f%%",mean(return_nasdaq))
p[[4]] = sprintf("STRATEGY Return %0.1f%%",mean(return_macd))

cat(paste0(unlist(p),collapse="\n"))
```

# Plot and Save Trades
```{r}
dir.create(file.path(Sys.getenv("USERPROFILE"),"Desktop","US Stock Trades"))

# plot and save all trades
for (i in 1:nrow(trade_hist1)) {

  # load stock
  path    = file.path(Sys.getenv("USERPROFILE"),"Desktop",paste0(trade_hist1$exchange[i]," stocks"))
  df      = read.csv(paste0(path,"/",trade_hist1$ticker[i],".csv"),sep=",",header=TRUE,stringsAsFactors=FALSE)
  df$Date = as.Date(df$Date)
  
  # crop df
  ema20       = ema(df$Close,20)
  ema50       = ema(df$Close,50)
  ema100      = ema(df$Close,100)
  ema200      = ema(df$Close,200)
  macd12_26_9 = macd(df$Close, 12, 26, 9)
  
  ema20  = ema20[df$Date > (trade_hist1$entry.date[i] - 25) & df$Date < (trade_hist1$exit.date[i] + 10)]
  ema50  = ema50[df$Date > (trade_hist1$entry.date[i] - 25) & df$Date < (trade_hist1$exit.date[i] + 10)]
  ema100 = ema100[df$Date > (trade_hist1$entry.date[i] - 25) & df$Date < (trade_hist1$exit.date[i] + 10)]
  ema200 = ema200[df$Date > (trade_hist1$entry.date[i] - 25) & df$Date < (trade_hist1$exit.date[i] + 10)]
  
  macd12_26_9$MACD   = macd12_26_9$MACD[df$Date > (trade_hist1$entry.date[i] - 25) & df$Date < (trade_hist1$exit.date[i] + 10)]
  macd12_26_9$signal = macd12_26_9$signal[df$Date > (trade_hist1$entry.date[i] - 25) & df$Date < (trade_hist1$exit.date[i] + 10)]
  
  df     = df[df$Date > (trade_hist1$entry.date[i] - 25) & df$Date < (trade_hist1$exit.date[i] + 10),]
  
  # plot
  savepath = file.path(Sys.getenv("USERPROFILE"),"Desktop","US Stock Trades")
  plot_stock(df,sub="macd",
             emas=list(ema20,ema50,ema100,ema200),
             macd=macd12_26_9,
             entry=trade_hist1$entry.date[i],exit=trade_hist1$exit.date[i],
             entry.price=trade_hist1$entry.price[i],exit.price=trade_hist1$exit.price[i],
             save=TRUE,savename=paste0(savepath,"/",i,"_",trade_hist1$ticker[i]," (",
                                       as.character(round(trade_hist1$outcome[i],1)),"%%)"))
  
}
```
